Change to separate lib directory hosting code
* This allows to load the code under test in isolation which avoids
  being forced to load rspec all the time when interacting with the
  code.
